requirements.txt
beautifulsoup4>=4.11.1
pytest>=7.0.0  # dev/test runner (recommended)
python-whois>=0.7.3
requests>=2.28.0

